requirements.txt
Flask==3.1.2
Flask-Cors==6.0.2
Flask-JWT-Extended==4.7.1
flask-sqlacodegen==2.0.0
flask_restful==0.3.10
Flask-Bcrypt==1.0.1
PyJWT==2.11.0
bcrypt==5.0.0
gunicorn==21.2.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.46
flask-marshmallow==1.4.0
marshmallow==4.2.2
marshmallow-sqlalchemy==1.4.2
Flask-Talisman==1.1.0
#gevent==24.2.1
pymysql==1.1.2
psycopg2-binary==2.9.11
python-dotenv==1.0.1